# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "TODO API",
        "title": "TODO API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/notes": {
            "get": {
                "description": "Find all notes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Find All Notes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.NoteResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Create Note",
                "parameters": [
                    {
                        "description": "New Note",
                        "name": "new_note",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create Note success",
                        "schema": {
                            "$ref": "#/definitions/handler.NoteResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notes/{id}": {
            "get": {
                "description": "Find Note by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Note"
                ],
                "summary": "Find Note by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note found",
                        "schema": {
                            "$ref": "#/definitions/handler.NoteResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateNoteRequest": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "handler.NoteResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}